chore: move schema to separate directory
Move the `new-transaction-schema` to lib/schema (was lib/types)
